requirements.txt
torch>=2.0.0
transformers>=4.40.0
datasets>=2.19.0
evaluate>=0.4.0
jiwer>=3.0.0
accelerate>=0.30.0
wandb>=0.17.0
soundfile>=0.12.1
librosa>=0.10.2
python-dotenv>=1.0.1
PyYAML>=6.0.1
matplotlib>=3.8.0
